+++ packages/graphiql/src/components/VariableEditor.tsx
@@ -31,6 +31,7 @@ type VariableEditorProps = {
   onMergeQuery: (value?: string) => void;
   onRunQuery: (value?: string) => void;
   editorTheme?: string;
+  active?: boolean;
 };
 
 /**
@@ -51,7 +52,7 @@ export class VariableEditor extends React.Component<VariableEditorProps> {
   editor: (CM.Editor & { options: any }) | null = null;
   cachedValue: string;
   private _node: HTMLElement | null = null;
-  ignoreChangeEvent: boolean;
+  ignoreChangeEvent: boolean = false;
   constructor(props: VariableEditorProps) {
     super(props);
 
@@ -59,7 +60,6 @@ export class VariableEditor extends React.Component<VariableEditorProps> {
     // editor is updated, which can later be used to protect the editor from
     // unnecessary updates during the update lifecycle.
     this.cachedValue = props.value || '';
-    this.ignoreChangeEvent = true;
   }
 
   componentDidMount() {
@@ -196,6 +196,12 @@ export class VariableEditor extends React.Component<VariableEditorProps> {
     return (
       <div
         className="codemirrorWrap"
+        // This horrible hack is necessary because a simple display none toggle
+        // causes one of the editors' gutters to break otherwise.
+        style={{
+          position: this.props.active ? 'relative' : 'absolute',
+          visibility: this.props.active ? 'visible' : 'hidden',
+        }}
         ref={node => {
           this._node = node as HTMLDivElement;
         }}