+++ packages/monaco-graphql/src/typings/monaco.d.ts
@@ -4,50 +4,34 @@
  *  This source code is licensed under the MIT license found in the
  *  LICENSE file in the root directory of this source tree.
  */
+
+// eslint-disable-next-line spaced-comment
+/// <reference types='monaco-editor'/>
+
 declare module monaco.languages.graphql {
+  import type {
+    SchemaLoader,
+    SchemaConfig as SchemaConfiguration,
+    SchemaResponse,
+  } from 'graphql-language-service';
+
+  import type { Options as PrettierConfig } from 'prettier';
+
+  import { MonacoGraphQLApi } from '../api';
+
   export interface IDisposable {
     dispose(): void;
   }
 
+  export type SchemaConfig = SchemaConfiguration;
+
   export interface IEvent<T> {
     (listener: (e: T) => any, thisArg?: any): IDisposable;
   }
-  export interface DiagnosticsOptions {
-    /**
-     * If set, the validator will be enabled and perform syntax validation as well as schema based validation.
-     */
-    readonly validate?: boolean;
-    /**
-     * If set, comments are tolerated. If set to false, syntax errors will be emitted for comments.
-     */
-    readonly allowComments?: boolean;
 
-    /**
-     * we'll just use this for now
-     */
-    readonly schemaUri: string;
-    /**
-     * A list of known schemas and/or associations of schemas to file names.
-     */
-    readonly schemas?: Array<{
-      /**
-       * The URI of the schema, which is also the identifier of the schema.
-       */
-      readonly uri?: string;
-      /**
-       * A list of file names that are associated to the schema. The '*' wildcard can be used. For example '*.schema.json', 'package.json'
-       */
-      readonly fileMatch?: string[];
-      /**
-       * The schema for the given URI.
-       */
-      readonly schema?: any;
-    }>;
-    /**
-     *  If set, the schema service would load schema content on-demand with 'fetch' if available
-     */
-    readonly enableSchemaRequest?: boolean;
-  }
+  export type FilePointer = string | string[];
+
+  export type FormattingOptions = PrettierConfig;
 
   export interface ModeConfiguration {
     /**
@@ -101,15 +85,27 @@ declare module monaco.languages.graphql {
     readonly selectionRanges?: boolean;
   }
 
+  export interface ICreateData {
+    languageId: string;
+    enableSchemaRequest: boolean;
+    schemaConfig: SchemaConfiguration;
+    schemaLoader: () => Promise<SchemaResponse>;
+    formattingOptions?: FormattingOptions;
+  }
+
   export interface LanguageServiceDefaults {
     readonly onDidChange: IEvent<LanguageServiceDefaults>;
-    readonly diagnosticsOptions: DiagnosticsOptions;
+    readonly schemaConfig: SchemaConfiguration;
+    readonly formattingOptions: FormattingOptions;
     readonly modeConfiguration: ModeConfiguration;
-    setDiagnosticsOptions(options: DiagnosticsOptions): void;
+    setSchemaConfig(options: SchemaConfiguration): void;
+    updateSchemaConfig(options: Partial<SchemaConfiguration>): void;
+    setSchemaUri(schemaUri: string): void;
+    setFormattingOptions(formattingOptions: FormattingOptions): void;
     setModeConfiguration(modeConfiguration: ModeConfiguration): void;
   }
 
-  // export const graphqlDefaults: LanguageServiceDefaults;
-}
+  export type api = MonacoGraphQLApi;
 
-// declare module ''monaco-editor-core/esm/vs/editor/editor.worker';
+  export const graphqlDefaults: LanguageServiceDefaults;
+}